Added fix to monitoring when data points are ignored when there are multiple stmts within one TX
