+++ core/src/main/java/org/polypheny/db/transaction/Statement.java
@@ -19,6 +19,7 @@ package org.polypheny.db.transaction;
 import org.polypheny.db.adapter.DataContext;
 import org.polypheny.db.information.InformationDuration;
 import org.polypheny.db.prepare.Context;
+import org.polypheny.db.monitoring.events.StatementEvent;
 import org.polypheny.db.processing.QueryProcessor;
 import org.polypheny.db.util.FileInputHandle;
 
@@ -37,6 +38,10 @@ public interface Statement {
     InformationDuration getRoutingDuration();
 
     InformationDuration getOverviewDuration();
+
+    StatementEvent getMonitoringEvent();
+
+    void setMonitoringEvent( StatementEvent event );
 
     void close();
