+++ core/src/main/java/org/polypheny/db/transaction/Transaction.java
@@ -16,6 +16,7 @@
 
 package org.polypheny.db.transaction;
 
+
 import java.util.List;
 import java.util.concurrent.atomic.AtomicBoolean;
 import org.polypheny.db.adapter.Adapter;
@@ -23,7 +24,6 @@ import org.polypheny.db.adapter.java.JavaTypeFactory;
 import org.polypheny.db.catalog.Catalog.QueryLanguage;
 import org.polypheny.db.catalog.entity.CatalogSchema;
 import org.polypheny.db.information.InformationManager;
-import org.polypheny.db.monitoring.events.StatementEvent;
 import org.polypheny.db.prepare.PolyphenyDbCatalogReader;
 import org.polypheny.db.processing.DataMigrator;
 import org.polypheny.db.processing.Processor;
@@ -73,10 +73,6 @@ public interface Transaction {
     long getNumberOfStatements();
 
     DataMigrator getDataMigrator();
-
-    StatementEvent getMonitoringEvent();
-
-    void setMonitoringEvent( StatementEvent event );
 
     void setUseCache( boolean useCache );
