+++ dbms/src/main/java/org/polypheny/db/routing/routers/DmlRouterImpl.java
@@ -596,8 +596,8 @@ public class DmlRouterImpl extends BaseRouter implements DmlRouter {
                         accessedPartitionList.add( identPart );
                     }
 
-                    if ( statement.getTransaction().getMonitoringEvent() != null ) {
-                        statement.getTransaction().getMonitoringEvent()
+                    if ( statement.getMonitoringEvent() != null ) {
+                        statement.getMonitoringEvent()
                                 .updateAccessedPartitions(
                                         Collections.singletonMap( catalogTable.id, accessedPartitionList )
                                 );
